requirements.txt
streamlit
chromadb
langchain
unstructured
sentence-transformers
openai
langchain-openai
langchain
pypdf
tiktoken
faiss-cpu
langchain-community



